requirements.txt
opencv-python
pillow~=11.3.0
pyzstd~=0.17.0
qrcode~=8.2
PyNaCl~=1.5.0